Catch permission errors for cross-domain iframes
In some circumstances (e.g, uploading files directly to Amazon S3)
we need to POST to another domain and can't guarantee we'll get
redirected back at the end of the request. In this case, it's
better to create an empty response object and let the calling code
decide how to handle it, rather than throwing a "permission denied"
error.